[ADD] made bulk add time persistent

+++ Phantom/Other/Helper.swift
@@ -147,6 +147,6 @@ struct Helper {
     
     static func getTimeOfDayString(_ timeOfDay: TimeInterval) -> String {
         let date = Date().startOfDay + timeOfDay
-        return DateFormatter.localizedString(from: date, dateStyle: .none, timeStyle: .long) // todo: short time
+        return DateFormatter.localizedString(from: date, dateStyle: .none, timeStyle: .short)
     }
 }
